+++ forest/cli/src/cli/mod.rs
@@ -6,6 +6,7 @@
 // check out the original commit history here:
 // https://github.com/ChainSafe/forest/commits/main/forest/src/cli/mod.rs
 
+mod attach_cmd;
 mod auth_cmd;
 mod chain_cmd;
 mod config_cmd;
@@ -31,10 +32,11 @@ use log::error;
 use serde::Serialize;
 
 pub(super) use self::{
-    auth_cmd::AuthCommands, chain_cmd::ChainCommands, config_cmd::ConfigCommands,
-    db_cmd::DBCommands, fetch_params_cmd::FetchCommands, mpool_cmd::MpoolCommands,
-    net_cmd::NetCommands, send_cmd::SendCommand, snapshot_cmd::SnapshotCommands,
-    state_cmd::StateCommands, sync_cmd::SyncCommands, wallet_cmd::WalletCommands,
+    attach_cmd::AttachCommand, auth_cmd::AuthCommands, chain_cmd::ChainCommands,
+    config_cmd::ConfigCommands, db_cmd::DBCommands, fetch_params_cmd::FetchCommands,
+    mpool_cmd::MpoolCommands, net_cmd::NetCommands, send_cmd::SendCommand,
+    snapshot_cmd::SnapshotCommands, state_cmd::StateCommands, sync_cmd::SyncCommands,
+    wallet_cmd::WalletCommands,
 };
 
 /// CLI structure generated when interacting with Forest binary
@@ -97,6 +99,9 @@ pub enum Subcommand {
     /// Database management
     #[command(subcommand)]
     DB(DBCommands),
+
+    /// Attach to daemon via a JavaScript console
+    Attach(AttachCommand),
 }
 
 /// Pretty-print a JSON-RPC error and exit