+++ node/rpc-client/src/chain_ops.rs
@@ -65,6 +65,9 @@ pub async fn chain_validate_tipset_checkpoints(
     call(CHAIN_VALIDATE_TIPSET_CHECKPOINTS, keys, auth_token).await
 }
 
-pub async fn chain_get_name(auth_token: &Option<String>) -> Result<ChainGetNameResult, Error> {
-    call(CHAIN_GET_NAME, (), auth_token).await
+pub async fn chain_get_name(
+    params: ChainGetNameParams,
+    auth_token: &Option<String>,
+) -> Result<ChainGetNameResult, Error> {
+    call(CHAIN_GET_NAME, params, auth_token).await
 }